# Make sure Send/Receive in StatsPusher can be canceled