--- a/core/services/synchronization/explorer_client_test.go
+++ b/core/services/synchronization/explorer_client_test.go
@@ -1,6 +1,7 @@
 package synchronization_test
 
 import (
+	"context"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
@@ -85,7 +86,7 @@ func TestWebSocketClient_Send_DefaultsToTextMessage(t *testing.T) {
 	defer explorerClient.Close()
 
 	expectation := `{"hello": "world"}`
-	explorerClient.Send([]byte(expectation))
+	explorerClient.Send(context.Background(), []byte(expectation))
 	cltest.CallbackOrTimeout(t, "receive stats", func() {
 		require.Equal(t, expectation, <-wsserver.ReceivedText)
 	})
@@ -100,7 +101,7 @@ func TestWebSocketClient_Send_TextMessage(t *testing.T) {
 	defer explorerClient.Close()
 
 	expectation := `{"hello": "world"}`
-	explorerClient.Send([]byte(expectation), synchronization.ExplorerTextMessage)
+	explorerClient.Send(context.Background(), []byte(expectation), synchronization.ExplorerTextMessage)
 	cltest.CallbackOrTimeout(t, "receive stats", func() {
 		require.Equal(t, expectation, <-wsserver.ReceivedText)
 	})
@@ -116,7 +117,7 @@ func TestWebSocketClient_Send_Binary(t *testing.T) {
 
 	address := common.HexToAddress("0xabc123")
 	addressBytes := address.Bytes()
-	explorerClient.Send(addressBytes, synchronization.ExplorerBinaryMessage)
+	explorerClient.Send(context.Background(), addressBytes, synchronization.ExplorerBinaryMessage)
 	cltest.CallbackOrTimeout(t, "receive stats", func() {
 		require.Equal(t, addressBytes, <-wsserver.ReceivedBinary)
 	})
@@ -131,7 +132,7 @@ func TestWebSocketClient_Send_Unsupported(t *testing.T) {
 	defer explorerClient.Close()
 
 	assert.PanicsWithValue(t, "send on explorer client received unsupported message type -1", func() {
-		explorerClient.Send([]byte(`{"hello": "world"}`), -1)
+		explorerClient.Send(context.Background(), []byte(`{"hello": "world"}`), -1)
 	})
 }
 
@@ -144,7 +145,7 @@ func TestWebSocketClient_Send_WithAck(t *testing.T) {
 	defer explorerClient.Close()
 
 	expectation := `{"hello": "world"}`
-	explorerClient.Send([]byte(expectation))
+	explorerClient.Send(context.Background(), []byte(expectation))
 	cltest.CallbackOrTimeout(t, "receive stats", func() {
 		require.Equal(t, expectation, <-wsserver.ReceivedText)
 		err := wsserver.Broadcast(`{"result": 200}`)
@@ -152,7 +153,7 @@ func TestWebSocketClient_Send_WithAck(t *testing.T) {
 	})
 
 	cltest.CallbackOrTimeout(t, "receive response", func() {
-		response, err := explorerClient.Receive()
+		response, err := explorerClient.Receive(context.Background())
 		assert.NoError(t, err)
 		assert.NotNil(t, response)
 	})
@@ -167,13 +168,13 @@ func TestWebSocketClient_Send_WithAckTimeout(t *testing.T) {
 	defer explorerClient.Close()
 
 	expectation := `{"hello": "world"}`
-	explorerClient.Send([]byte(expectation))
+	explorerClient.Send(context.Background(), []byte(expectation))
 	cltest.CallbackOrTimeout(t, "receive stats", func() {
 		require.Equal(t, expectation, <-wsserver.ReceivedText)
 	})
 
 	cltest.CallbackOrTimeout(t, "receive response", func() {
-		_, err := explorerClient.Receive(100 * time.Millisecond)
+		_, err := explorerClient.Receive(context.Background(), 100*time.Millisecond)
 		assert.Error(t, err)
 		assert.Equal(t, err, synchronization.ErrReceiveTimeout)
 	}, 300*time.Millisecond)