The StatsPusher tests are very slow, this should speed them up a bit...

Before:
ok  	github.com/smartcontractkit/chainlink/core/services/synchronization	79.264s

After: (ctx first step)
ok  	github.com/smartcontractkit/chainlink/core/services/synchronization	52.237s

Before: -run Auth
ok      github.com/smartcontractkit/chainlink/core/services/synchronization     45.389s

After: (Start() optimization) -run Auth
ok      github.com/smartcontractkit/chainlink/core/services/synchronization     0.386s

final: 
ok  	github.com/smartcontractkit/chainlink/core/services/synchronization	8.971s